feat: object <-> array mapper implementation
As part of an upcoming refactor, this patch provides a mapper implementation for converting objects to array serializations, and vice versa.
Serializations MUST include a "__type" key referencing the class they were serialized from.
This approach allows decoupling tasks from this library, and puts the onus on consumers to shape how serialization and hydration occur.

An `EmptyObjectMapper` implementation provides an easy way to provide a mapper for an empty event class.