refactor: move JSON serialization/deserialization to Mapper, and map tasks in RedisTaskQueue
Updates the code to move all serialization/deserialization to the Mapper, and renames methods to reflect this:

- `extract(): array` becomes `toString(): string`
- `hydrate(array): object` becomes `toObject(string): object`
- `canHydrate(array): bool` becomes `canCastToObject(string): bool`

The RedisTaskQueue was updated to reflect these changes, which eliminated a number of validation checks.
Addtionally, `retrieveNextTask()` was updated to return the deserialized task instance, and all `retrieve*Tasks()` methods were updated to return arrays of tasks, instead of serialized JSON.

These changes eliminated the need for the `Worker` class, as all work to pull the task and cast it can be done via the queue itself.